# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Luma Restaurant REST API
servers:
  - url: /api
paths:
...
  /reservations:
    get:
      operationId: listReservations
      responses: { "200": { description: Reservations } }
    post:
      operationId: createReservation
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/ReservationInput" } }
      responses: { "201": { description: Created } }
  /reservations/{id}:
    get:
      operationId: getReservation
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { description: Reservation } }
    put:
      operationId: updateReservation
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/ReservationUpdate" } }
      responses: { "200": { description: Updated } }
    delete:
      operationId: deleteReservation
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "204": { description: Deleted } }
  /orders:
    get:
      operationId: listOrders
      responses: { "200": { description: Orders } }
    post:
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/OrderInput" } }
      responses: { "201": { description: Created } }
  /orders/{id}:
    get:
      operationId: getOrder
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "200": { description: Order } }
    put:
      operationId: updateOrder
      parameters: [{ $ref: "#/components/parameters/Id" }]
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: "#/components/schemas/OrderUpdate" } }
      responses: { "200": { description: Updated } }
    delete:
      operationId: deleteOrder
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses: { "204": { description: Deleted } }
...
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
  schemas:
...
    ReservationInput:
      type: object
      required: [customerName, email, phone, reservationDate, reservationTime, guests]
      properties: { customerName: { type: string }, email: { type: string, format: email }, phone: { type: string }, reservationDate: { type: string, format: date }, reservationTime: { type: string }, guests: { type: integer, minimum: 1, maximum: 50 }, specialRequest: { type: string } }
    ReservationUpdate:
      type: object
      properties: { status: { type: string, enum: [pending, confirmed, cancelled, completed] }, customerName: { type: string }, email: { type: string, format: email }, phone: { type: string }, reservationDate: { type: string, format: date }, reservationTime: { type: string }, guests: { type: integer }, specialRequest: { type: string } }
    OrderInput:
      type: object
      required: [customerName, customerEmail, customerPhone, orderType, items]
      properties: { customerName: { type: string }, customerEmail: { type: string, format: email }, customerPhone: { type: string }, orderType: { type: string, enum: [dine_in, takeaway, delivery] }, deliveryAddress: { type: string }, items: { type: array, items: { type: object, required: [menuItemId, quantity], properties: { menuItemId: { type: integer }, quantity: { type: integer, minimum: 1 } } } } }
    OrderUpdate:
      type: object
      properties: { status: { type: string, enum: [pending, confirmed, preparing, ready, out_for_delivery, delivered, cancelled] } }
...