# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luma Restaurant REST API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputItemsItem } from './orderInputItemsItem';
import type { OrderInputOrderType } from './orderInputOrderType';

export interface OrderInput {
  customerName: string;
  customerEmail: string;
  customerPhone: string;
  orderType: OrderInputOrderType;
  deliveryAddress?: string;
  items: OrderInputItemsItem[];
}