# lib/api-zod/src/generated/types/reservationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luma Restaurant REST API
 * OpenAPI spec version: 0.1.0
 */

export interface ReservationInput {
  customerName: string;
  email: string;
  phone: string;
  reservationDate: Date;
  reservationTime: string;
  /**
     * @minimum 1
     * @maximum 50
     */
  guests: number;
  specialRequest?: string;
}