# lib/api-zod/src/generated/types/reservationUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Luma Restaurant REST API
 * OpenAPI spec version: 0.1.0
 */
import type { ReservationUpdateStatus } from './reservationUpdateStatus';

export interface ReservationUpdate {
  status?: ReservationUpdateStatus;
  customerName?: string;
  email?: string;
  phone?: string;
  reservationDate?: Date;
  reservationTime?: string;
  guests?: number;
  specialRequest?: string;
}